Remove unnecessary code, various rename
